feat: Refactor VSCode extension build process and enhance command registration
