+++ jac/support/vscode_ext/jac/CHANGELOG.md
@@ -5,5 +5,3 @@ All notable changes to the "jac" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-
-- Initial release
