+++ jac/support/vscode_ext/jac/src/constants.ts
@@ -1,10 +1,1 @@
 export const TERMINAL_NAME = "Jac Terminal";
-
-export const COMMANDS = {
-    RUN_FILE: 'jaclang-extension.runCurrentFile',
-    CHECK_FILE: 'jaclang-extension.checkCurrentFile',
-    SERVE_FILE: 'jaclang-extension.serveCurrentFile',
-    SELECT_ENV: 'jaclang-extension.selectEnv',
-    GET_JAC_PATH: 'extension.jaclang-extension.getJacPath',
-    VISUALIZE: 'jac.visualize'
-};
