+++ jac/support/vscode_ext/jac/src/test/extension.test.ts
@@ -5,6 +5,11 @@ import * as sinon from 'sinon';
 suite('Extension Tests', () => {
     suiteSetup(async () => {
         const ext = vscode.extensions.getExtension('jaseci-labs.jaclang-extension');
+        console.log('Extension:', ext);
+        if (typeof ext?.activate === 'function') {
+            console.log('It has an activate method.');
+        }
+
         if (ext) {
             await ext.activate();
         }
@@ -13,6 +18,9 @@ suite('Extension Tests', () => {
     test('Commands are registered', async () => {
         const commands = await vscode.commands.getCommands(true);
         console.log('Registered commands:', commands);
+        // print all commands start in 'jaclang-extension'
+        const jacCommands = commands.filter(cmd => cmd.startsWith('jaclang-extension.'));
+        console.log('Jac commands:', jacCommands);
         assert.ok(commands.includes('jaclang-extension.selectEnv'), 'jaclang-extension.selectEnv not registered');
         assert.ok(commands.includes('jaclang-extension.runCurrentFile'), 'runCurrentFile not registered');
         assert.ok(commands.includes('jaclang-extension.checkCurrentFile'), 'checkCurrentFile not registered');
