+++ jac/support/vscode_ext/jac/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/constants.ts","./src/extension.ts","./src/utils.ts","./src/commands/index.ts","./src/environement/manager.ts","./src/lsp/client.ts","./src/test/commands.test.ts","./src/test/extension.test.ts","./src/test/lspClient.test.ts","./src/test/runTest.ts","./src/test/suite/index.ts","./src/utils/envDetection.ts","./src/utils/jacRunner.ts","./src/visual_debugger/visdbg.ts","./src/webview/visualDebugger.ts"],"version":"5.8.3"}
+{"root":["./src/constants.ts","./src/extension.ts","./src/utils.ts","./src/commands/commands.ts","./src/commands/constants.ts","./src/commands/index.ts","./src/environement/manager.ts","./src/lsp/client.ts","./src/test/commands.test.ts","./src/test/extension.test.ts","./src/test/lspClient.test.ts","./src/test/runTest.ts","./src/test/suite/index.ts","./src/utils/envDetection.ts","./src/utils/jacRunner.ts","./src/visual_debugger/visdbg.ts","./src/webview/visualDebugger.ts"],"version":"5.8.3"}
